# Fix std.hash benchmarks