@jedisct1 2023-05-31
@matu3ba It's hashing different, unpredictable inputs.

In release mode, results from the previous benchmark code were completely off. The total number of bytes used to compute the throughput was not the real one.