@tiehuis 2023-06-02
Good catch. The same issue exists in the crypto benchmarks, too. However it looks as though the reverse is somewhat true there, and we were undercalculating. We were passing blocks of digest_length and so were often doing partial rounds/lots of memcpy where we don't need to. This is most noticeable on the crypto functions which don't use the sponge-like absorb process.

## Before

```
              md5:        588 MiB/s
             sha1:        818 MiB/s
           sha256:       1475 MiB/s
           sha512:        584 MiB/s
         sha3-256:        521 MiB/s
         sha3-512:        313 MiB/s
        shake-128:        680 MiB/s
        shake-256:        588 MiB/s
   turboshake-128:       1190 MiB/s
   turboshake-256:       1091 MiB/s
          blake2s:        537 MiB/s
          blake2b:        796 MiB/s
           blake3:        932 MiB/s
```

## After

```
              md5:        817 MiB/s
             sha1:        978 MiB/s
           sha256:       1954 MiB/s
           sha512:        650 MiB/s
         sha3-256:        633 MiB/s
         sha3-512:        336 MiB/s
        shake-128:        776 MiB/s
        shake-256:        633 MiB/s
   turboshake-128:       1492 MiB/s
   turboshake-256:       1217 MiB/s
          blake2s:        503 MiB/s
          blake2b:        848 MiB/s
           blake3:       1053 MiB/s
```

---

```diff
 lib/std/crypto/benchmark.zig | 30 +++++++++++++++++-------------
 1 file changed, 17 insertions(+), 13 deletions(-)

diff --git a/lib/std/crypto/benchmark.zig b/lib/std/crypto/benchmark.zig
index 696a9c310..a89bbb084 100644
--- a/lib/std/crypto/benchmark.zig
+++ b/lib/std/crypto/benchmark.zig
@@ -10,6 +10,8 @@ const crypto = std.crypto;
 const KiB = 1024;
 const MiB = 1024 * KiB;
 
+const block_size = 8 * 8192;
+
 var prng = std.rand.DefaultPrng.init(0);
 const random = prng.random();
 
@@ -34,17 +36,18 @@ const hashes = [_]Crypto{
     Crypto{ .ty = crypto.hash.Blake3, .name = "blake3" },
 };
 
-pub fn benchmarkHash(comptime Hash: anytype, comptime bytes: comptime_int) !u64 {
-    var h = Hash.init(.{});
+pub fn benchmarkHash(comptime Hash: anytype, comptime bytes: comptime_int, allocator: std.mem.Allocator) !u64 {
+    var blocks = try allocator.alloc(u8, bytes);
+    defer allocator.free(blocks);
+    random.bytes(blocks);
 
-    var block: [Hash.digest_length]u8 = undefined;
-    random.bytes(block[0..]);
+    var h = Hash.init(.{});
 
     var offset: usize = 0;
     var timer = try Timer.start();
     const start = timer.lap();
-    while (offset < bytes) : (offset += block.len) {
-        h.update(block[0..]);
+    while (offset < bytes) : (offset += block_size) {
+        h.update(blocks[offset..][0..block_size]);
     }
     mem.doNotOptimizeAway(&h);
     const end = timer.read();
@@ -72,9 +75,10 @@ const macs = [_]Crypto{
     Crypto{ .ty = crypto.auth.cmac.CmacAes128, .name = "aes-cmac" },
 };
 
-pub fn benchmarkMac(comptime Mac: anytype, comptime bytes: comptime_int) !u64 {
-    var in: [512 * KiB]u8 = undefined;
-    random.bytes(in[0..]);
+pub fn benchmarkMac(comptime Mac: anytype, comptime bytes: comptime_int, allocator: std.mem.Allocator) !u64 {
+    var blocks = try allocator.alloc(u8, bytes);
+    defer allocator.free(blocks);
+    random.bytes(blocks);
 
     const key_length = if (Mac.key_length == 0) 32 else Mac.key_length;
     var key: [key_length]u8 = undefined;
@@ -84,8 +88,8 @@ pub fn benchmarkMac(comptime Mac: anytype, comptime bytes: comptime_int) !u64 {
     var offset: usize = 0;
     var timer = try Timer.start();
     const start = timer.lap();
-    while (offset < bytes) : (offset += in.len) {
-        Mac.create(mac[0..], in[0..], key[0..]);
+    while (offset < bytes) : (offset += block_size) {
+        Mac.create(mac[0..], blocks[offset..][0..block_size], key[0..]);
         mem.doNotOptimizeAway(&mac);
     }
     const end = timer.read();
@@ -482,14 +486,14 @@ pub fn main() !void {
 
     inline for (hashes) |H| {
         if (filter == null or std.mem.indexOf(u8, H.name, filter.?) != null) {
-            const throughput = try benchmarkHash(H.ty, mode(128 * MiB));
+            const throughput = try benchmarkHash(H.ty, mode(128 * MiB), arena_allocator);
             try stdout.print("{s:>17}: {:10} MiB/s\n", .{ H.name, throughput / (1 * MiB) });
         }
     }
 
     inline for (macs) |M| {
         if (filter == null or std.mem.indexOf(u8, M.name, filter.?) != null) {
-            const throughput = try benchmarkMac(M.ty, mode(128 * MiB));
+            const throughput = try benchmarkMac(M.ty, mode(128 * MiB), arena_allocator);
             try stdout.print("{s:>17}: {:10} MiB/s\n", .{ M.name, throughput / (1 * MiB) });
         }
     }
```