@jedisct1 2023-06-02
@tiehuis Indeed, results for hash functions in the crypto benchmark were also flawed.
They have been updated them as well. Thanks!